requirements.txt
asgiref==3.2.6
Django==3.0.4
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
project-rename==0.5
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
